Make newly_added_tracks also return playlist_tracks
I realized that both newly_added_tracks and push_new_tracks were calling
get_playlist_tracks on the same playlist, so that's 2 calls to the API
of the same thing. The thing is that newly_added_tracks was doing it
directly to the API and not through datapipe because it needed to check
the time added as it was going through the tracks. So what I did was
have it make a list of playlist tracks too as it was doing its own thing
with checking added by time, that way it was still only one call to the
API while newly_added_tracks still being able to check the time added.